A08_Music_Catalog_API-DRF-08_project/apps/catalog/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import ArtistViewSet, AlbumViewSet, SongViewSet, AlbumSongViewSet

router = DefaultRouter()
router.register(r'artists', ArtistViewSet, basename='artist')
router.register(r'albums', AlbumViewSet, basename='album')
router.register(r'songs', SongViewSet, basename='song')
router.register(r'album-songs', AlbumSongViewSet, basename='album-song')

urlpatterns = [
    path('', include(router.urls)),
]
