A08_Music_Catalog_API-DRF-08_project/requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.10.6
cryptography==43.0.3
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.1.3
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-decouple==2.1
django-environ==0.12.0
django-filter==25.1
django-jazzmin==3.0.1
django-jet==1.0.8
django-phonenumber-field==8.1.0
django-redis==6.0.0
django-simple-history==3.8.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
dnspython==2.7.0
docopt==0.6.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
drf-yasg==1.21.5
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
fonttools==4.56.0
frozenlist==1.7.0
greenlet==3.2.4
gTTS==2.5.4
gunicorn==20.1.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
lxml==5.3.1
magic-filter==1.0.12
Mako==1.3.10
Markdown==3.7
MarkupSafe==3.0.2
multidict==6.6.4
oauthlib==3.2.2
openapi-codec==1.3.2
oscrypto==1.3.0
packaging==24.2
passlib==1.7.4
phonenumbers==9.0.13
pillow==11.0.0
pipreqs==0.4.13
pluggy==1.6.0
propcache==0.3.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.2
pyHanko==0.26.0
pyhanko-certvalidator==0.26.5
PyJWT==2.10.0
pypdf==5.3.1
pyphen==0.17.2
pytest==8.4.1
pytest-django==4.11.1
python-bidi==0.6.6
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.5.0
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
redis==4.4.0
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.27.0
rsa==4.9.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
SQLAlchemy==2.0.43
sqlparse==0.4.2
starlette==0.47.2
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.3
uvicorn==0.35.0
weasyprint==64.1
webencodings==0.5.1
whitenoise==6.4.0
xhtml2pdf==0.2.17
yarg==0.1.10
yarl==1.20.1
zopfli==0.2.3.post1
